module-lattice: remove `Flatten` and `Unflatten`
They're now in `hybrid-array` as of v0.4.7
