+++ module-lattice/src/algebra.rs
@@ -1,4 +1,4 @@
-use super::utils::Truncate;
+use super::truncate::Truncate;
 
 use array::{Array, ArraySize, typenum::U256};
 use core::fmt::Debug;
@@ -80,7 +80,7 @@ macro_rules! define_field {
                 let product = x * Self::BARRETT_MULTIPLIER;
                 let quotient = product >> Self::BARRETT_SHIFT;
                 let remainder = x - quotient * Self::QLL;
-                Self::small_reduce($crate::utils::Truncate::truncate(remainder))
+                Self::small_reduce($crate::truncate::Truncate::truncate(remainder))
             }
         }
     };
