--- module-lattice/src/utils.rs
@@ -1,177 +1,0 @@
-use array::{
-    Array, ArraySize,
-    typenum::{Prod, Quot, U0, Unsigned},
-};
-use core::{
-    mem::ManuallyDrop,
-    ops::{Div, Mul, Rem},
-    ptr,
-};
-
-/// Safely truncate an unsigned integer value to shorter representation
-pub trait Truncate<T> {
-    /// Truncate value to the width of `Self`.
-    fn truncate(x: T) -> Self;
-}
-
-macro_rules! define_truncate {
-    ($from:ident, $to:ident) => {
-        impl Truncate<$from> for $to {
-            // Truncation should always function as intended here:
-            // - we ensure `$to` is small enough to infallibly convert to `$from` via the
-            //   `$from::from($to::MAX)` conversion, which will fail if not widening.
-            // - we are deliberately masking to the smaller size, i.e. truncation is intentional
-            //   (though that's not enough for `clippy` for some reason). Arguably the truncation
-            //   of the `as` cast is sufficient, but this makes it explicit
-            #[allow(clippy::cast_possible_truncation)]
-            fn truncate(x: $from) -> $to {
-                (x & $from::from($to::MAX)) as $to
-            }
-        }
-    };
-}
-
-define_truncate!(u32, u16);
-define_truncate!(u64, u16);
-define_truncate!(u64, u32);
-define_truncate!(u128, u8);
-define_truncate!(u128, u16);
-define_truncate!(u128, u32);
-define_truncate!(usize, u8);
-define_truncate!(usize, u16);
-
-/// Defines a sequence of sequences that can be merged into a bigger overall sequence.
-pub trait Flatten<T, M: ArraySize> {
-    /// Size of the output array.
-    type OutputSize: ArraySize;
-
-    /// Flatten array.
-    fn flatten(self) -> Array<T, Self::OutputSize>;
-}
-
-impl<T, N, M> Flatten<T, Prod<M, N>> for Array<Array<T, M>, N>
-where
-    N: ArraySize,
-    M: ArraySize + Mul<N>,
-    Prod<M, N>: ArraySize,
-{
-    type OutputSize = Prod<M, N>;
-
-    fn flatten(self) -> Array<T, Self::OutputSize> {
-        let whole = ManuallyDrop::new(self);
-
-        // SAFETY: this is the reverse transmute between [T; K*N] and [[T; K], M], which is guaranteed
-        // to be safe by the Rust memory layout of these types.
-        #[allow(unsafe_code)]
-        unsafe {
-            ptr::read(whole.as_ptr().cast())
-        }
-    }
-}
-
-/// Defines a sequence that can be split into a sequence of smaller sequences of uniform size.
-pub trait Unflatten<M>
-where
-    M: ArraySize,
-{
-    /// Part of the array we're decomposing into.
-    type Part;
-
-    /// Unflatten array into `Self::Part` chunks.
-    fn unflatten(self) -> Array<Self::Part, M>;
-}
-
-impl<T, N, M> Unflatten<M> for Array<T, N>
-where
-    N: ArraySize + Div<M> + Rem<M, Output = U0>,
-    M: ArraySize,
-    Quot<N, M>: ArraySize,
-{
-    type Part = Array<T, Quot<N, M>>;
-
-    fn unflatten(self) -> Array<Self::Part, M> {
-        let part_size = Quot::<N, M>::USIZE;
-        let whole = ManuallyDrop::new(self);
-
-        // SAFETY: this is doing the same thing as what is done in `Array::split`.
-        // Basically, this is doing transmute between [T; K*N] and [[T; K], M], which is guaranteed to
-        // be safe by the Rust memory layout of these types.
-        #[allow(unsafe_code)]
-        Array::from_fn(|i| unsafe {
-            let offset = i.checked_mul(part_size).expect("overflow");
-            ptr::read(whole.as_ptr().add(offset).cast())
-        })
-    }
-}
-
-impl<'a, T, N, M> Unflatten<M> for &'a Array<T, N>
-where
-    N: ArraySize + Div<M> + Rem<M, Output = U0>,
-    M: ArraySize,
-    Quot<N, M>: ArraySize,
-{
-    type Part = &'a Array<T, Quot<N, M>>;
-
-    fn unflatten(self) -> Array<Self::Part, M> {
-        let part_size = Quot::<N, M>::USIZE;
-        let mut ptr: *const T = self.as_ptr();
-
-        // SAFETY: this is doing the same thing as what is done in `Array::split`.
-        // Basically, this is doing transmute between [T; K*N] and [[T; K], M], which is guaranteed to
-        // be safe by the Rust memory layout of these types.
-        #[allow(unsafe_code)]
-        Array::from_fn(|_i| unsafe {
-            let part = &*(ptr.cast());
-            ptr = ptr.add(part_size);
-            part
-        })
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use array::{
-        Array,
-        sizes::{U2, U5},
-    };
-
-    #[test]
-    fn flatten() {
-        let flat: Array<u8, _> = Array([1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
-        let unflat2: Array<Array<u8, _>, _> = Array([
-            Array([1, 2]),
-            Array([3, 4]),
-            Array([5, 6]),
-            Array([7, 8]),
-            Array([9, 10]),
-        ]);
-        let unflat5: Array<Array<u8, _>, _> =
-            Array([Array([1, 2, 3, 4, 5]), Array([6, 7, 8, 9, 10])]);
-
-        // Flatten
-        let actual = unflat2.flatten();
-        assert_eq!(flat, actual);
-
-        let actual = unflat5.flatten();
-        assert_eq!(flat, actual);
-
-        // Unflatten
-        let actual: Array<Array<u8, U2>, U5> = flat.unflatten();
-        assert_eq!(unflat2, actual);
-
-        let actual: Array<Array<u8, U5>, U2> = flat.unflatten();
-        assert_eq!(unflat5, actual);
-
-        // Unflatten on references
-        let actual: Array<&Array<u8, U2>, U5> = (&flat).unflatten();
-        for (i, part) in actual.iter().enumerate() {
-            assert_eq!(&unflat2[i], *part);
-        }
-
-        let actual: Array<&Array<u8, U5>, U2> = (&flat).unflatten();
-        for (i, part) in actual.iter().enumerate() {
-            assert_eq!(&unflat5[i], *part);
-        }
-    }
-}
